+++ aiku/aiku-main/src/main/java/aiku_main/controller/ScheduleController.java
@@ -1,9 +1,7 @@
 package aiku_main.controller;
 
 import aiku_main.dto.*;
-import aiku_main.dto.schedule.ScheduleAddDto;
-import aiku_main.dto.schedule.ScheduleUpdateDto;
-import aiku_main.dto.team.TeamScheduleListResDto;
+import aiku_main.dto.schedule.*;
 import aiku_main.service.ScheduleService;
 import common.response.BaseResponse;
 import common.response.BaseResultDto;
