+++ aiku/aiku-main/src/main/java/aiku_main/dto/schedule/MemberScheduleListEachResDto.java
@@ -1,5 +1,6 @@
-package aiku_main.dto;
+package aiku_main.dto.schedule;
 
+import aiku_main.dto.LocationDto;
 import com.querydsl.core.annotations.QueryProjection;
 import common.domain.ExecStatus;
 import lombok.Getter;
