+++ aiku/aiku-main/src/main/java/aiku_main/dto/schedule/MemberScheduleListResDto.java
@@ -1,4 +1,4 @@
-package aiku_main.dto;
+package aiku_main.dto.schedule;
 
 import lombok.Getter;
