+++ aiku/aiku-main/src/main/java/aiku_main/repository/ScheduleQueryRepositoryCustomImpl.java
@@ -2,7 +2,9 @@ package aiku_main.repository;
 
 import aiku_main.application_event.domain.ScheduleArrivalMember;
 import aiku_main.dto.*;
-import aiku_main.dto.team.TeamScheduleListEachResDto;
+import aiku_main.dto.schedule.MemberScheduleListEachResDto;
+import aiku_main.dto.schedule.ScheduleMemberResDto;
+import aiku_main.dto.schedule.TeamScheduleListEachResDto;
 import com.querydsl.core.types.Projections;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.core.types.dsl.Expressions;
@@ -10,7 +12,6 @@ import com.querydsl.jpa.JPAExpressions;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import common.domain.ExecStatus;
 import common.domain.schedule.QScheduleMember;
-import common.domain.schedule.QScheduleResult;
 import common.domain.schedule.ScheduleMember;
 import common.domain.schedule.ScheduleResult;
 import lombok.RequiredArgsConstructor;
