+++ aiku/aiku-main/src/main/java/aiku_main/service/ScheduleService.java
@@ -5,10 +5,7 @@ import aiku_main.application_event.domain.ScheduleArrivalResult;
 import aiku_main.application_event.publisher.PointChangeEventPublisher;
 import aiku_main.application_event.publisher.ScheduleEventPublisher;
 import aiku_main.dto.*;
-import aiku_main.dto.schedule.ScheduleAddDto;
-import aiku_main.dto.schedule.ScheduleUpdateDto;
-import aiku_main.dto.team.TeamScheduleListEachResDto;
-import aiku_main.dto.team.TeamScheduleListResDto;
+import aiku_main.dto.schedule.*;
 import aiku_main.exception.ScheduleException;
 import aiku_main.exception.TeamException;
 import aiku_main.kafka.KafkaProducerService;
