+++ aiku/aiku-main/src/test/java/aiku_main/service/ScheduleServiceTest.java
@@ -2,7 +2,7 @@ package aiku_main.service;
 
 import aiku_main.application_event.publisher.ScheduleEventPublisher;
 import aiku_main.dto.LocationDto;
-import aiku_main.dto.ScheduleDetailResDto;
+import aiku_main.dto.schedule.ScheduleDetailResDto;
 import aiku_main.dto.schedule.ScheduleUpdateDto;
 import aiku_main.kafka.KafkaProducerService;
 import aiku_main.repository.MemberRepository;
